requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
celery==4.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1.7
django-ckeditor==6.0.0
django-debug-toolbar==3.2
django-elasticsearch-dsl==7.1.4
django-extensions==3.1.2
django-js-asset==1.2.2
djangorestframework==3.12.4
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
idna==2.10
install==1.3.4
jsonfield==3.1.0
kombu==4.6.11
nepali==0.4.2
# Pillow==8.2.0
prompt-toolkit==3.0.19
# psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.14.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
